@@ -26,9 +26,10 @@
     },
     "dependencies": {
         "@asgardeo/auth-spa": "^0.4.5",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-json-view": "^1.21.3"
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-json-view": "^1.21.3",
+        "react-notifications-component": "^4.0.1"
     },
     "devDependencies": {
         "@babel/core": "^7.17.5",
@@ -36,12 +37,12 @@
         "@babel/preset-react": "^7.16.7",
         "babel-loader": "^8.2.3",
         "css-loader": "^6.6.0",
-        "html-loader": "^3.1.0",
+        "html-loader": "^4.1.0",
         "html-webpack-plugin": "^5.5.0",
         "style-loader": "^3.3.1",
         "url-loader": "^4.1.1",
-        "webpack": "^5.69.1",
-        "webpack-cli": "^4.9.2",
-        "webpack-dev-server": "^4.7.4"
+        "webpack": "^5.74.0",
+        "webpack-cli": "^4.10.0",
+        "webpack-dev-server": "^4.9.3"
     }
 }
